// src/laudo/dto/create-laudo.dto.ts
export class CreateLaudoDto {
  data_inicio: string;
  data_fim: string;
  cnes: string;
  cidade?: string;
  numero_processo?: string;
  data_distribuicao?: string;
  data_citacao?: string;
  data_fim_correcao?: string;
  nome_fantasia?: string;
  razao_social?: string;
  cnpj?: string;
  ivr_tunep?: string;
}